apps/api/api/requirements.txt
azure-functions
azure-functions-worker
requests
python-dotenv